+++ src/api/IBMi.ts
@@ -3,7 +3,7 @@ import * as vscode from "vscode";
 import * as node_ssh from "node-ssh";
 import { ConnectionConfiguration } from "./Configuration";
 
-import Tools from './Tools';
+import {Tools} from './Tools';
 import path from 'path';
 
 let remoteApps = [
@@ -500,7 +500,7 @@ export default class IBMi {
 
             if (output.stdout) {
               const [row] = Tools.db2Parse(output.stdout);
-              if (row && row.CHARACTER_CODE_SET_ID !== `null`) {
+              if (row && row.CHARACTER_CODE_SET_ID !== `null` && typeof row.CHARACTER_CODE_SET_ID === 'number') {
                 this.qccsid = row.CHARACTER_CODE_SET_ID;
               }
             }
@@ -514,8 +514,8 @@ export default class IBMi {
 
               if (output.stdout) {
                 const rows = Tools.db2Parse(output.stdout);
-                const ccsid = rows.find((row: any) => row.SYSTEM_VALUE_NAME === `QCCSID`);
-                if (ccsid) {
+                const ccsid = rows.find(row => row.SYSTEM_VALUE_NAME === `QCCSID`);
+                if (ccsid && typeof ccsid.CURRENT_NUMERIC_VALUE === 'number') {
                   this.qccsid = ccsid.CURRENT_NUMERIC_VALUE;
                 }
               }
@@ -543,7 +543,7 @@ export default class IBMi {
             });
             if (output.stdout) {
               const [row] = Tools.db2Parse(output.stdout);
-              if (row && row.LOCAL !== `null`) {
+              if (row && row.LOCAL !== `null` && typeof row.LOCAL === 'string') {
                 this.variantChars.local = row.LOCAL;
               }
             } else {
