+++ src/api/IBMiContent.js
@@ -5,7 +5,7 @@ const util = require(`util`);
 let fs = require(`fs`);
 const tmp = require(`tmp`);
 const csv = require(`csv/sync`);
-const Tools = require(`./Tools`);
+const {Tools} = require(`./Tools`);
 const {ObjectTypes} = require(`../schemas/Objects`);
 
 const tmpFile = util.promisify(tmp.file);
@@ -97,7 +97,7 @@ module.exports = class IBMiContent {
     spf = spf.toUpperCase();
     mbr = mbr.toUpperCase();
 
-    const path = Tools.qualifyPath(asp, lib, spf, mbr);
+    const path = Tools.qualifyPath(lib, spf, mbr, asp);
     const tempRmt = this.ibmi.getTempRemote(path);
     const tmpobj = await tmpFile();
     const client = this.ibmi.client;
@@ -148,7 +148,7 @@ module.exports = class IBMiContent {
     mbr = mbr.toUpperCase();
 
     const client = this.ibmi.client;
-    const path = Tools.qualifyPath(asp, lib, spf, mbr);
+    const path = Tools.qualifyPath(lib, spf, mbr, asp);
     const tempRmt = this.ibmi.getTempRemote(path);
     const tmpobj = await tmpFile();
 
@@ -170,7 +170,7 @@ module.exports = class IBMiContent {
   /**
    * Run an SQL statement
    * @param {string} statement
-   * @returns {Promise<any[]>} Result set
+   * @returns {Promise<Tools.DB2Row[]>} Result set
    */
   async runSQL(statement) {
     const { 'QZDFMDB2.PGM': QZDFMDB2 } = this.ibmi.remoteFeatures;
@@ -218,7 +218,7 @@ module.exports = class IBMiContent {
       return data;
 
     } else {
-      const tempRmt = this.ibmi.getTempRemote(Tools.qualifyPath(undefined, lib, file, mbr));
+      const tempRmt = this.ibmi.getTempRemote(Tools.qualifyPath(lib, file, mbr));
 
       await this.ibmi.remoteCommand(
         `QSYS/CPYTOIMPF FROMFILE(` +
