+++ src/api/Tools.ts
@@ -1,0 +1,138 @@
+export namespace Tools {
+  export interface DB2Headers {
+    name: string
+    from: number
+    length: number
+  }
+
+  export interface DB2Row extends Record<string, string | number | null> {
+
+  }
+  
+  /**
+   * Parse standard out for `/usr/bin/db2`
+   * @param output /usr/bin/db2's output
+   * @returns rows
+   */
+  export function db2Parse(output : string) : DB2Row[]{
+    let gotHeaders = false;
+    let figuredLengths = false;
+    let iiErrorMessage = false;
+
+    let data = output.split(`\n`);
+
+    let headers : DB2Headers[];
+
+    let SQLSTATE;
+
+    const rows : DB2Row[] = [];
+
+    data.forEach((line, index) => {
+      const trimmed = line.trim();
+      if (trimmed.length === 0 && iiErrorMessage) iiErrorMessage = false;
+      if (trimmed.length === 0 || index === data.length - 1) return;
+      if (trimmed === `DB2>`) return;
+      if (trimmed.startsWith(`DB20`)) return; // Notice messages
+      if (trimmed === `?>`) return;
+
+      if (trimmed === `**** CLI ERROR *****`) {
+        iiErrorMessage = true;
+        if (data.length > index + 3) {
+          SQLSTATE = data[index + 1].trim();
+
+          if (SQLSTATE.includes(`:`)) {
+            SQLSTATE = SQLSTATE.split(`:`)[1].trim();
+          }
+
+          if (!SQLSTATE.startsWith(`01`)) {
+            throw new Error(`${data[index + 3]} (${SQLSTATE})`);
+          }
+        }
+        return;
+      }
+
+      if (iiErrorMessage) return;
+
+      if (gotHeaders === false) {
+        headers = line.split(` `)
+          .filter(header => header.length > 0)
+          .map(header => {
+            return {
+              name: header,
+              from: 0,
+              length: 0,
+            };
+        });
+
+        gotHeaders = true;
+      } else if (figuredLengths === false) {
+        let base = 0;
+        line.split(` `).forEach((header, index) => {
+          headers[index].from = base;
+          headers[index].length = header.length;
+
+          base += header.length + 1;
+        });
+
+        figuredLengths = true;
+      } else {
+        let row : DB2Row = {};
+
+        headers.forEach(header => {
+          const strValue = line.substring(header.from, header.from + header.length).trimEnd();
+
+          let realValue: string | number | null = strValue;
+
+          // is value a number?
+          if (strValue.startsWith(` `)) {
+            const asNumber = Number(strValue.trim());
+            if (!isNaN(asNumber)) {
+              realValue = asNumber;
+            }
+          } else if (strValue === `-`) {
+            realValue = null; //null?
+          }
+
+          row[header.name] = realValue;
+        });
+
+        rows.push(row);
+      }
+    });
+
+    return rows;
+  }
+
+  export function makeid() {
+    let text = `O_`;
+    const possible =
+      `ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789`;
+
+    for (let i = 0; i < 8; i++)
+      text += possible.charAt(Math.floor(Math.random() * possible.length));
+
+    return text;
+  }
+
+  /**
+   * Build the IFS path string to a member
+   * @param library
+   * @param object
+   * @param member
+   * @param iasp Optional: an iASP name
+   */
+   export function qualifyPath(library: string, object: string, member: string, iasp?: string) {
+    const path =
+      (iasp && iasp.length > 0 ? `/${iasp}` : ``) + `/QSYS.lib/${library}.lib/${object}.file/${member}.mbr`;
+    return path;
+  }
+
+  /**
+   * @param Path
+   * @returns the escaped path
+   */
+   export function escapePath(Path: string): string {
+    const path = Path.replace(/'|"|\$|\\| /g, matched => `\\`.concat(matched));
+    return path;
+  }
+}
